Create README.md
Processing of MRI images is one of the part of this field. Identification tumour are rising problem as there is rise in people being affected by tumour this rise is caused by lot of factors from habits to pollution. Locating a tumour as always been problem as that requires a lot of experience of human anatomy which is requires a lot of time. This project describes the proposed strategy to detect & extraction of brain tumour from patients. Detection and extraction of tumour from MRI scan images of the brain is done by using MATLAB software.  Executing the code  Open the Code in matlab Change the directory for each input image Example for image 5 I=imread('C:\Users\eminovicm\Desktop\y(1).jpg');  Run the code